Removed add_done_callback and switched AsCompleted->as_completed.
The reason for removing `add_done_callback` is that it really did very little, and was easily expressed using in-line wrapper coroutines instead. One anyway already had to define the callback, so this seems simpler overall.

We switched from `iterator = AsCompleted(...)` to `iterator = yield as_completed(...)` to reflect the more natural place to schedule the coroutines - when the generator is created, not on its first `.get()`.